Refactored routes into resources and controllers
