+++ resources/scoreController.js
