+++ resources/trialController.js
